PRMDR-80: Migrate patient search lambda and project refactor (#13)
* Migrate patient search lambda added
* Project structure refactor
* GitHub actions modified to work with make commands
* GitHub actions for lambda functions have been modified to include search Lambda

Note: Due to the change of the project structure, the CI action will not trigger for this build. As such, a manual deployment to ndr-dev has been performed, which is showing as successful here https://github.com/nhsconnect/national-document-repository/actions/runs/5955511355. Future builds should be detected once the changed actions have been applied to the main branch.



---------

Co-authored-by: Scott Alexander <scott.alexander@madetech.com>