Set title to optional
title in Attachment had been set to remove the default value. There is a
slight difference, title: Optional[str] still requires a value of None
if no value is passed in. title: Optional[str = None is fully optional
in that if no value is passed it will put a default.